requirements.txt
click
python-chess
